PA2: submitting the form. This now fails because of CSRF protection return page prefills the 'zoobar' and 'to' fields of transfer form
When submitting the form the POST will fail since it does not have the
secret CSRF token. However since the form target is the doubly nested
frame then the response will be loaded to that frame (without allowing
it to framebust). The response comes back with the prefilled 'to' and
'zoobars' fields that we submitted. Now all we need to do is cover up
the zoobar UI and put a facade over it to invite the victim to click.